adaptive_istio_rate_limit/requirements.txt
# TrendMaster-AI Adaptive Istio Rate Limiting System
# Production Dependencies

# Core Python packages
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<2.0.0
scipy>=1.7.0,<2.0.0

# Machine Learning and Forecasting
prophet>=1.1.0,<2.0.0
scikit-learn>=1.0.0,<2.0.0

# HTTP and API clients
requests>=2.25.0,<3.0.0
urllib3>=1.26.0,<2.0.0

# YAML and configuration
PyYAML>=6.0,<7.0
pydantic>=1.8.0,<2.0.0

# Kubernetes integration
kubernetes>=18.20.0,<25.0.0
pykube-ng>=21.3.0,<23.0.0

# Date and time handling
python-dateutil>=2.8.0,<3.0.0
pytz>=2021.1

# Logging and monitoring
structlog>=21.1.0,<23.0.0
prometheus-client>=0.11.0,<1.0.0

# Data validation and serialization
marshmallow>=3.12.0,<4.0.0
jsonschema>=3.2.0,<5.0.0

# CLI and argument parsing
click>=8.0.0,<9.0.0
rich>=10.0.0,<13.0.0

# Environment and configuration
python-dotenv>=0.19.0,<1.0.0

# Networking and async
aiohttp>=3.7.0,<4.0.0
asyncio-throttle>=1.0.0,<2.0.0

# Data processing utilities
tabulate>=0.8.0,<1.0.0
jinja2>=3.0.0,<4.0.0

# Development and testing dependencies (optional)
# Uncomment for development environment
# pytest>=6.2.0,<8.0.0
# pytest-asyncio>=0.15.0,<1.0.0
# pytest-cov>=2.12.0,<4.0.0
# black>=21.6.0,<23.0.0
# flake8>=3.9.0,<5.0.0
# mypy>=0.910,<1.0.0
# pre-commit>=2.13.0,<3.0.0
# bandit>=1.7.0,<2.0.0
# safety>=1.10.0,<3.0.0

# Production monitoring and observability
opencensus-ext-prometheus>=0.7.1,<1.0.0
opencensus-ext-logging>=0.1.0,<1.0.0

# Security and authentication
cryptography>=3.4.0,<40.0.0
PyJWT>=2.1.0,<3.0.0

# Performance and optimization
cachetools>=4.2.0,<6.0.0
memory-profiler>=0.58.0,<1.0.0

# Data formats and serialization
orjson>=3.6.0,<4.0.0
msgpack>=1.0.0,<2.0.0

# System utilities
psutil>=5.8.0,<6.0.0
distro>=1.6.0,<2.0.0

# Networking utilities for Prometheus queries
httpx>=0.24.0,<1.0.0
tenacity>=8.0.0,<9.0.0

# Mathematical and statistical functions
statsmodels>=0.12.0,<1.0.0

# Time series analysis
tslearn>=0.5.0,<1.0.0

# Configuration management
omegaconf>=2.1.0,<3.0.0
hydra-core>=1.1.0,<2.0.0

# Async utilities
asyncio-mqtt>=0.10.0,<1.0.0

# Production WSGI server (if needed for web interface)
gunicorn>=20.1.0,<21.0.0
uvicorn>=0.15.0,<1.0.0

# Health checks and monitoring
healthcheck>=1.3.0,<2.0.0

# Backup and archiving
tarfile-stream>=0.1.0,<1.0.0

# Version constraints for security
setuptools>=65.5.1  # Security fix
wheel>=0.38.0       # Security fix
pip>=22.3.0         # Security fix

# Platform-specific dependencies
# Linux-specific packages
# systemd-python>=234; sys_platform == "linux"

# macOS-specific packages  
# pyobjc-framework-SystemConfiguration>=7.0; sys_platform == "darwin"

# Windows-specific packages
# pywin32>=227; sys_platform == "win32"
